+++ spark/src/main/scala/ai/chronon/spark/LoggingSchema.scala
@@ -1,6 +1,7 @@
-package ai.chronon.online
+package ai.chronon.spark
 
 import ai.chronon.api.{HashUtils, StructField, StructType}
+import ai.chronon.online.{AvroCodec, JoinCodec}
 import com.google.gson.Gson
 
 import scala.util.ScalaJavaConversions.MapOps
